# frontend/src/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "战鸽数据",
    "description": "Zhange Stats · Steam 游玩统计与圈子成员管理",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/taygedo/status": {
      "get": {
        "tags": [
          "taygedo"
        ],
        "summary": "Taygedo Status",
        "operationId": "taygedo_status_api_taygedo_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "include_roles",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Roles"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaygedoStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/taygedo/logs": {
      "get": {
        "tags": [
          "taygedo"
        ],
        "summary": "Taygedo Logs",
        "description": "兼容占位：历史列表已弃用；今日状态见 status（读 *_checkin_logs 缓存）。",
        "operationId": "taygedo_logs_api_taygedo_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CheckinLogOut"
                  },
                  "title": "Response Taygedo Logs Api Taygedo Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/taygedo/bind/password": {
      "post": {
        "tags": [
          "taygedo"
        ],
        "summary": "Taygedo Bind Password",
        "operationId": "taygedo_bind_password_api_taygedo_bind_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaygedoBindPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaygedoStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/taygedo/bind/sms/send": {
      "post": {
        "tags": [
          "taygedo"
        ],
        "summary": "Taygedo Bind Sms Send",
        "operationId": "taygedo_bind_sms_send_api_taygedo_bind_sms_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaygedoBindSmsSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaygedoBindSmsSendResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/taygedo/bind/sms": {
      "post": {
        "tags": [
          "taygedo"
        ],
        "summary": "Taygedo Bind Sms",
        "operationId": "taygedo_bind_sms_api_taygedo_bind_sms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaygedoBindSmsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaygedoStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/taygedo/bind/json": {
      "post": {
        "tags": [
          "taygedo"
        ],
        "summary": "Taygedo Bind Json",
        "operationId": "taygedo_bind_json_api_taygedo_bind_json_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaygedoBindJsonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaygedoStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/taygedo/bind": {
      "delete": {
        "tags": [
          "taygedo"
        ],
        "summary": "Taygedo Unbind",
        "operationId": "taygedo_unbind_api_taygedo_bind_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaygedoStatusOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "taygedo"
        ],
        "summary": "Taygedo Update Bind",
        "operationId": "taygedo_update_bind_api_taygedo_bind_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaygedoBindUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaygedoStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/taygedo/checkin": {
      "post": {
        "tags": [
          "taygedo"
        ],
        "summary": "Taygedo Checkin Now",
        "operationId": "taygedo_checkin_now_api_taygedo_checkin_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckinLogOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "game_code": {
            "type": "string",
            "title": "Game Code"
          },
          "game_name": {
            "type": "string",
            "title": "Game Name"
          },
          "role_uid": {
            "type": "string",
            "title": "Role Uid"
          },
          "role_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Name"
          },
          "channel_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_label": {
            "type": "string",
            "title": "Status Label",
            "default": ""
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "awards_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awards Text"
          },
          "checkin_date": {
            "type": "string",
            "title": "Checkin Date"
          },
          "checked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Checked At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "game_code",
          "game_name",
          "role_uid",
          "status",
          "checkin_date",
          "checked_at"
        ],
        "title": "CheckinLogOut",
        "description": "签到记录公共结构（森空岛 / 塔吉多等共用）。"
      },
...
      "CheckinResponse": {
        "properties": {
          "skipped": {
            "type": "boolean",
            "title": "Skipped",
            "default": false
          },
          "ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ok"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/CheckinResultItem"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "CheckinResponse"
      },
      "CheckinResultItem": {
        "properties": {
          "game_code": {
            "type": "string",
            "title": "Game Code"
          },
          "game_name": {
            "type": "string",
            "title": "Game Name"
          },
          "role_uid": {
            "type": "string",
            "title": "Role Uid"
          },
          "role_name": {
            "type": "string",
            "title": "Role Name"
          },
          "channel_name": {
            "type": "string",
            "title": "Channel Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_label": {
            "type": "string",
            "title": "Status Label",
            "default": ""
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "awards_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awards Text"
          },
          "extra_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Text"
          }
        },
        "type": "object",
        "required": [
          "game_code",
          "game_name",
          "role_uid",
          "role_name",
          "channel_name",
          "status",
          "message"
        ],
        "title": "CheckinResultItem"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaygedoBindJsonRequest": {
        "properties": {
          "credentials_json": {
            "type": "string",
            "maxLength": 8192,
            "minLength": 8,
            "title": "Credentials Json"
          }
        },
        "type": "object",
        "required": [
          "credentials_json"
        ],
        "title": "TaygedoBindJsonRequest"
      },
      "TaygedoBindPasswordRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 6,
            "title": "Phone"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 4,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "password"
        ],
        "title": "TaygedoBindPasswordRequest"
      },
      "TaygedoBindSmsRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 6,
            "title": "Phone"
          },
          "captcha": {
            "type": "string",
            "maxLength": 8,
            "minLength": 4,
            "title": "Captcha"
          },
          "device_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 8,
            "title": "Device Id"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "captcha",
          "device_id"
        ],
        "title": "TaygedoBindSmsRequest"
      },
      "TaygedoBindSmsSendRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 6,
            "title": "Phone"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          }
        },
        "type": "object",
        "required": [
          "phone"
        ],
        "title": "TaygedoBindSmsSendRequest"
      },
      "TaygedoBindSmsSendResponse": {
        "properties": {
          "device_id": {
            "type": "string",
            "title": "Device Id"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "验证码已发送"
          }
        },
        "type": "object",
        "required": [
          "device_id"
        ],
        "title": "TaygedoBindSmsSendResponse"
      },
      "TaygedoBindUpdate": {
        "properties": {
          "auto_checkin": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Checkin"
          },
          "checkin_hour": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 23.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkin Hour"
          },
          "checkin_minute": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 59.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkin Minute"
          }
        },
        "type": "object",
        "title": "TaygedoBindUpdate"
      },
      "TaygedoRoleOut": {
        "properties": {
          "game_code": {
            "type": "string",
            "title": "Game Code"
          },
          "game_name": {
            "type": "string",
            "title": "Game Name"
          },
          "uid": {
            "type": "string",
            "title": "Uid"
          },
          "role_name": {
            "type": "string",
            "title": "Role Name"
          },
          "channel_name": {
            "type": "string",
            "title": "Channel Name"
          }
        },
        "type": "object",
        "required": [
          "game_code",
          "game_name",
          "uid",
          "role_name",
          "channel_name"
        ],
        "title": "TaygedoRoleOut"
      },
      "TaygedoStatusOut": {
        "properties": {
          "bound": {
            "type": "boolean",
            "title": "Bound"
          },
          "auto_checkin": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Checkin"
          },
          "checkin_hour": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkin Hour"
          },
          "checkin_minute": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkin Minute"
          },
          "phone_mask": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Mask"
          },
          "bound_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bound At"
          },
          "last_checkin_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checkin At"
          },
          "last_checkin_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checkin Date"
          },
          "last_checkin_ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checkin Ok"
          },
          "last_checkin_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checkin Summary"
          },
          "token_ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Ok"
          },
          "token_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Error"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/TaygedoRoleOut"
            },
            "type": "array",
            "title": "Roles"
          },
          "today_results": {
            "items": {
              "$ref": "#/components/schemas/CheckinResultItem"
            },
            "type": "array",
            "title": "Today Results"
          },
          "today_logs": {
            "items": {
              "$ref": "#/components/schemas/CheckinLogOut"
            },
            "type": "array",
            "title": "Today Logs"
          }
        },
        "type": "object",
        "required": [
          "bound"
        ],
        "title": "TaygedoStatusOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}